fix: autentication

+++ src/api/routes.py
@@ -13,8 +13,6 @@ from flask_bcrypt import Bcrypt
 from flask_mail import Mail, Message
 
 
-# TODO: Pasar token cada vez que un user o trainer quiera ver algo 
-
 api = Blueprint('api', __name__)
 CORS(api)  # Allow CORS requests to this API
 bcrypt = Bcrypt()
@@ -128,7 +126,6 @@ def handle_signup_trainer():
         return response_body, 200
 
 
-
 @api.route('/administrators', methods=['GET'])
 @jwt_required()
 def handle_admins():
@@ -185,7 +182,6 @@ def handle_admin_signup():
         return response_body, 200
 
 
-# TODO: Hacer que el admin pueda crear specializaciones, aplicar token, para que las puedan ver entrenadores y admin
 @api.route('/specializations', methods=["POST"])
 @jwt_required()
 def add_specializations():
@@ -206,8 +202,10 @@ def add_specializations():
         response_body["message"] = "Specialization already exists"
         return response_body, 400
     new_specialization = Specializations(name=data["name"].lower(), 
-                                            description=data.get("description"), 
-                                            logo_url=data.get("logo_url"))
+                                         description=data.get("description"), 
+                                         logo_url=data.get("logo_url"))
+                                         description=data.get("description"), 
+                                         logo_url=data.get("logo_url"))
     db.session.add(new_specialization)
     db.session.commit()
     response_body["message"] = "Specialization created"
@@ -228,7 +226,6 @@ def get_specializations():
     return response_body, 200
 
 
-# Endpoints unico para los login
 @api.route('/login/<user_type>', methods=['POST'])
 def handle_login(user_type):
     response_body = {}
@@ -255,7 +252,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {user.email}'
             return response_body, 401
         access_token = create_access_token(identity={"user": user.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": user.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"user": user.serialize(), 
                                     "role": user_type}
@@ -271,7 +269,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {trainer.email}'
             return response_body, 401
         access_token = create_access_token(identity={"trainer": trainer.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": trainer.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"trainer": trainer.serialize(), 
                                     "role": user_type}
@@ -287,7 +286,8 @@ def handle_login(user_type):
             response_body['message'] = f'Wrong password for email {administrator.email}'
             return response_body, 401
         access_token = create_access_token(identity={"administrator": administrator.email,
-                                                     "role": user_type})
+                                                     "role": user_type,
+                                                     "id": administrator.id})
         response_body['message'] = 'Successfully logged in!'
         response_body['results'] = {"administrator": administrator.serialize(), 
                                     "role": user_type}
@@ -429,6 +429,10 @@ def handle_user_classes(id):
     user = Users.query.get(id)
     classes_user = UsersClasses.query.filter_by(user_id=id).all()
     classes_trainer = TrainersClasses.query.get(id)
+    current_user = get_jwt_identity()
+    if current_user['role'] not in ['administrators', 'users'] or current_user['id'] != user.id:
+        response_body["message"] = 'Not allowed!'
+        return response_body, 405
     if not user:
         response_body["message"] = "User not found"
         return response_body, 404
@@ -455,14 +459,12 @@ def handle_user_classes(id):
         if not trainer_class:
             response_body["message"] = "No Trainer class available"
             return response_body, 404
-        new_class = UsersClasses(
-            amount=data["amount"], 
-            stripe_status="Cart", 
-            trainer_status="Pending", 
-            value=0,
-            user_id=id,
-            class_id=data["class_id"]
-        )
+        new_class = UsersClasses(amount=data["amount"], 
+                                 stripe_status= "Cart", 
+                                 trainer_status= "Pending", 
+                                 value=0,
+                                 user_id=id,
+                                 class_id=data["class_id"])
         db.session.add(new_class)
         db.session.commit()
         response_body["message"] = "Class added"
@@ -476,6 +478,7 @@ def handle_trainer_classes(id):
     response_body = {}
     trainer = Trainers.query.get(id)
     classes_trainer = TrainersClasses.query.filter_by(trainer_id=id).all()
+    current_user = get_jwt_identity()
     if not trainer:
         response_body["message"] = "Trainer not found"
         return response_body, 404
@@ -487,6 +490,9 @@ def handle_trainer_classes(id):
         response_body["Trainer classes"] = [class_trainer.serialize() for class_trainer in classes_trainer]
         return response_body, 200
     if request.method == "POST":
+        if (current_user['role'] == 'trainers' and current_user['id'] != trainer.id) or (current_user["role"] != "administrators"):
+            response_body["message"] = 'Not allowed!'
+            return response_body, 405
         data = request.json
         if not data:
             response_body["message"] = "No data provided"
@@ -495,16 +501,14 @@ def handle_trainer_classes(id):
         if existing_class:
             response_body["message"] = "Trainer class already exists for this datetime"
             return response_body, 400
-        new_trainer_class = TrainersClasses(
-            trainer_id=id, 
-            address=data["address"], 
-            capacity=data["capacity"], 
-            duration=data["duration"],
-            date=data["date"],
-            price=data["price"],
-            training_type=data["training_type"],
-            training_level=data["training_level"]
-        )
+        new_trainer_class = TrainersClasses(trainer_id=id, 
+                                            address=data["address"], 
+                                            capacity=data["capacity"], 
+                                            duration=data["duration"],
+                                            date=data["date"],
+                                            price=data["price"],
+                                            training_type=data["training_type"],
+                                            training_level=data["training_level"])
         db.session.add(new_trainer_class)
         db.session.commit()
         response_body["message"] = "New class create"
@@ -518,6 +522,7 @@ def handle_trainer_class(id, class_id):
     response_body = {}
     trainer = Trainers.query.get(id)
     trainer_class = TrainersClasses.query.filter(TrainersClasses.trainer_id == id, TrainersClasses.id == class_id).first()
+    current_user = get_jwt_identity()
     if not trainer: 
         response_body["message"] = "Trainer not found"
         return response_body, 404
@@ -528,35 +533,38 @@ def handle_trainer_class(id, class_id):
         response_body["message"] = "Trainer class"
         response_body["class"] = trainer_class.serialize()
         return response_body, 200
-    if request.method == "DELETE":
-        # No hace cancelar las classes si hay usuarios apuntados a ella
-        classes_user = UsersClasses.query.filter_by(class_id=class_id).all()
-        if classes_user:
-            response_body["message"] = "Unable to delete class, it has associated users"
-            return response_body, 400
-        db.session.delete(trainer_class)
-        db.session.commit()
-        response_body["message"] = "Class deleted successfully"
-        response_body["class"] = trainer_class.serialize()
-        return response_body, 200
-    if request.method == "PATCH":
-        data = request.json
-        if not data:
-            response_body["message"] = "No data provided for update"
-            return response_body, 400
-        if 'address' in data:
-            trainer_class.address = data["address"]
-        if 'duration' in data:
-            trainer_class.duration = data["duration"]
-        if 'date' in data:
-            trainer_class.date = data["date"]
-        if 'price' in data:
-            trainer_class.price = data["price"]
-        db.session.add(trainer_class)
-        db.session.commit()
-        response_body["message"] = "Class update"
-        response_body["result"] = trainer_class.serialize()
-        return response_body, 200
+    if (current_user['role'] == 'trainers' and current_user['id'] == trainer.id) or (current_user["role"] == "administrators"):
+        if request.method == "DELETE":
+            # No hace cancelar las classes si hay usuarios apuntados a ella
+            classes_user = UsersClasses.query.filter_by(class_id=class_id).all()
+            if classes_user:
+                response_body["message"] = "Unable to delete class, it has associated users"
+                return response_body, 400
+            db.session.delete(trainer_class)
+            db.session.commit()
+            response_body["message"] = "Class deleted successfully"
+            response_body["class"] = trainer_class.serialize()
+            return response_body, 200
+        if request.method == "PATCH":
+            data = request.json
+            if not data:
+                response_body["message"] = "No data provided for update"
+                return response_body, 400
+            if 'address' in data:
+                trainer_class.address = data["address"]
+            if 'duration' in data:
+                trainer_class.duration = data["duration"]
+            if 'date' in data:
+                trainer_class.date = data["date"]
+            if 'price' in data:
+                trainer_class.price = data["price"]
+            db.session.add(trainer_class)
+            db.session.commit()
+            response_body["message"] = "Class update"
+            response_body["result"] = trainer_class.serialize()
+            return response_body, 200
+    response_body["message"] = 'Not allowed!'
+    return response_body, 405
             
 
 @api.route('/users/<int:id>/classes/<int:class_id>', methods=["GET", "DELETE"])
@@ -569,6 +577,10 @@ def handle_user_class(id, class_id):
     user_class = UsersClasses.query.filter_by(user_id=id, class_id=class_id).first()
     # Buscar la clase del entrenador por ID de clase
     trainer_class = TrainersClasses.query.get(class_id)
+    current_user = get_jwt_identity()
+    if (current_user['role'] == 'users' and current_user['id'] != user.id) or (current_user["role"] != "administrators"):
+            response_body["message"] = 'Not allowed!'
+            return response_body, 405
     if not user:
         response_body["message"] = "User not found"
         return response_body, 404
@@ -590,7 +602,6 @@ def handle_user_class(id, class_id):
         return response_body, 200
 
 
-# New endpoint to GET and CREATE TrainersSpecializations
 @api.route('/trainers/<int:id>/specializations', methods=['GET','POST'])
 @jwt_required()
 def handle_trainers_specializations(id):
@@ -603,7 +614,7 @@ def handle_trainers_specializations(id):
         trainers_specializations = db.session.query(TrainersSpecializations).filter_by(trainer_id = id).all()
         if not trainers_specializations:
             response_body['message'] = 'No trainer specializations for trainer id ' + str(id)
-            return response_body,404
+            return response_body, 404
         response_body['message'] = 'Trainer specializations for trainer ' + str(id)
         response_body['results'] = [spec.serialize() for spec in trainers_specializations]
         return response_body,200
@@ -616,14 +627,13 @@ def handle_trainers_specializations(id):
         if not specialization:
             response_body['message'] = 'No specialization with id ' + data['specialization_id'] + ' !'
             return response_body, 404
-        new_trainer_specialization = TrainersSpecializations(
-            certification = data['certification'],
-            status = "Requested",
-            specialization_id = data['specialization_id'],  
-            trainer_id = id
-        )
+        new_trainer_specialization = TrainersSpecializations(certification = data['certification'],
+                                                             status = "Requested",
+                                                             specialization_id = data['specialization_id'],  
+                                                             trainer_id = id)
         db.session.add(new_trainer_specialization)
         db.session.commit()
         response_body['message'] = 'New specialization connected with trainer ' + str(id)
         response_body['results'] = new_trainer_specialization.serialize()
         return response_body,201
+
